Modified an operation to convert fee values type into double

+++ src/main/scala/it/unica/blockchain/analyses/ethereum/mongo/TxWithFees.scala
@@ -6,6 +6,10 @@ import it.unica.blockchain.db.DatabaseSettings
 import it.unica.blockchain.mongo.Collection
 import it.unica.blockchain.externaldata.rates.EthereumRates
 
+/**This analysis uses external data.
+  * Make sure you have installed all the required libraries!
+  * Checkout the README file */
+
 object TxWithFees {
   def main(args: Array[String]): Unit = {
     val blockchain = BlockchainLib.getEthereumBlockchain(new EthereumSettings("http://localhost:8545"))
@@ -27,7 +31,7 @@ object TxWithFees {
           ("value", tx.value.doubleValue()/weiIntoEth.doubleValue()),
           ("creates", creates),
           ("gas", tx.gas),
-          ("fee", (tx.gas * tx.gasPrice)/weiIntoEth),
+          ("fee", (tx.gas * tx.gasPrice).doubleValue()/weiIntoEth.doubleValue()),
           ("rate", EthereumRates.getRate(block.date))
         )
         txWithFees.append(list)
